requirements.txt
django
pandas
scikit-learn==1.6.1
numpy
xgboost
shap
joblib
gunicorn
gevent
whitenoise
azure-storage-blob>=12.21.0
azure-identity>=1.17.0
python-dotenv>=1.0.1